Rework tooling API project configuration progress event generation to be more consistent with generation of test and file download progress events.
